requirements.txt
telethon
pyyaml
pil
pytesseract
BeautifulSoup